Be hoang

+++ BE/Common/src/main/java/com/pbl5/common/entity/Shop.java
@@ -28,11 +28,12 @@ public class Shop extends IdBaseEntity{
     private String address;
 
     private String photo;
-    private Integer rating;
+    private Float rating;
     private Integer productAmount;
     private Integer peopleTracking;
 
     @OneToOne
     @JoinColumn(name = "user_id")
     private User user;
+
 }
